SCL-SEM/Contrastive Learning/requirements.txt
numpy==1.23.3
Pillow==9.4.0
PyYAML==6.0
torch==1.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.1
